refactor(visualization): move reset map to default into global settings
The destructive 'Reset map to default' action lived in the floating
view-cube toolbar as an icon-only button, making accidental clicks easy.
Move it into the Global Configuration dialog next to 'Reset global
settings', restyle it as a labelled danger button, and keep the
confirmation dialog. Relocate the reset button, confirm dialog and
MapResetStore from the viewCubeToolbox feature into globalSettings.
