+++ visualization/app/codeCharta/features/globalSettings/components/globalConfigurationDialog/resetMapButton/resetMapButton.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, viewChild } from "@angular/core"
 import { ConfirmResetMapDialogComponent } from "../confirmResetMapDialog/confirmResetMapDialog.component"
 
 @Component({
-    selector: "cc-toolbox-reset-map-button",
+    selector: "cc-reset-map-button",
     templateUrl: "./resetMapButton.component.html",
     imports: [ConfirmResetMapDialogComponent],
     changeDetection: ChangeDetectionStrategy.OnPush
