+++ visualization/app/codeCharta/features/globalSettings/stores/mapReset.store.ts
