+++ visualization/app/codeCharta/features/viewCubeToolbox/components/viewCubeToolbox/viewCubeToolbox.component.html
@@ -2,5 +2,4 @@
     <cc-toolbox-center-map-button></cc-toolbox-center-map-button>
     <cc-toolbox-screenshot-button></cc-toolbox-screenshot-button>
     <cc-toolbox-presentation-mode-button></cc-toolbox-presentation-mode-button>
-    <cc-toolbox-reset-map-button></cc-toolbox-reset-map-button>
 </div>
